Commit by Sachin and Anush together.
Note - It is the combined work of both Sachin and Anush and both have contributed equally.

Both of us sat together and implemented new features such as breakfast, lunch, and dinner pages.
Changed the menu.
Added time features to check when the forms will open.
Added date and day detection feature to show that day's mess menu only.
Added feature to check that the paricular form can be filled only once in that day.
Added dynamic form for different items in menu.
Added popups and alerts to tell the user that he cannot fill the form twice or before time.
Added the feature to give the total summation of the calorie intake of the food items selectd by the user.
Adder error handling features to check for errors in API calls.
Fixed some previous bugs also.